requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
sqlalchemy==2.0.36
alembic==1.14.0
aiohttp==3.10.11
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
apscheduler==3.10.4
python-dateutil==2.9.0
